pyproject.toml
[project]
name = "tib-mcp"
version = "0.1.0"
description = "TIB ran MCP server for all scholarly tools"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.0",
    "python-dotenv>=1.1",
    "requests>=2.33.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
    "pytest-cov>=6.0",
    "ruff>=0.11",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
